refactor(modularization): Phase 4 Step 4-S2b — governance DB split (ApiKey + ApiKeyStatus + AuditLog + AuditResource → aperag/domains/governance/db/models.py) (#25)
Second sub-step of Phase 4 Step 4-S2.

**New module** ``aperag/domains/governance/db/models.py``:
- ``ApiKey`` + ``AuditLog`` class bodies moved verbatim from the
  legacy aggregate; paired enums (``ApiKeyStatus`` +
  ``AuditResource``) moved alongside them.
- Local ``_random_id`` / ``_enum_column`` helpers (per-domain DB
  module pattern; Phase 6 cleanup consolidates).

**aperag/db/models.py**:
- Class bodies replaced with breadcrumb comments.
- Re-export block extended with governance symbols (bottom-of-file
  where no other class in the module binds these enums — unlike
  Role which had to go top for ``Invitation``).

**Verification**:
- ``api_key`` / ``audit_log`` tables still registered on
  ``Base.metadata``.
- Boundary + Phase 3 re-export audit gates 17/17 green.
- ``ruff check`` + ``ruff format --check`` clean.
